backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.0.0
pydantic==2.10.5
numpy==1.24.3
